pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-rag-ru"
version = "0.1.0"
description = "Local Russian-first RAG Web Service"
readme = "rus_rag_local_architecture.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "numpy>=1.26",
  "sentence-transformers>=5.3.0",
  "transformers==4.51.0",
  "faiss-cpu>=1.8.0",
  "requests>=2.32.0",
  "pypdf>=4.2.0",
  "python-docx>=1.1.2",
  "einops>=0.8.2",
  "datasets>=2.20.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "PyJWT>=2.9.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1.4",
  "email-validator>=2.2.0",
]

[tool.setuptools]
packages = ["rag", "app", "app.api", "app.api.routes", "app.core", "app.db", "app.models", "app.schemas", "app.services"]
